Extract BuildCacheClientModule
